requirements.txt
django==4.1.7
asgiref==3.6.0
gunicorn==20.1.0
psycopg2==2.9.5
sqlparse==0.4.3
whitenoise==6.4.0
python-dotenv==1.0.0
requests==2.28.0
django-sesame==3.1
websockets==10.4
osu.py==1.1.0
mysql-connector-python==8.0.33
schedule==1.2.1